Reporting
All of the reporting code put in new Reporting.cs
