+++ WpfCSCS.csproj
@@ -162,6 +162,7 @@
     <Compile Include="AdictionaryLocal\SY_INDEXES.cs" />
     <Compile Include="AdictionaryLocal\SY_TABLES.cs" />
     <Compile Include="Btrieve.cs" />
+    <Compile Include="Reporting.cs" />
     <Compile Include="SpecialWindow.cs" />
     <Compile Include="App.xaml.cs">
       <DependentUpon>App.xaml</DependentUpon>
